requirements.txt
google_auth_oauthlib==0.2.0
requests==2.18.4
PyDrive==1.3.1
urlextract==0.9
protobuf==3.7.0
google_api_python_client==1.7.8
